+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [DEV] - 2020-02-06
+## [2.0.0] - 2020-02-17
 * added RebuildSearchIndexCommand
 * added disable search index option
 * [BC BREAK] Renamed bundle class